@@ -1,24 +1,23 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFrameworks>net8.0;net6.0;netstandard2.1</TargetFrameworks>
+		<TargetFrameworks>net9.0;net8.0</TargetFrameworks>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<LangVersion>Latest</LangVersion>
+		<Nullable>enable</Nullable>
 		<Description>Server Redis Caching Library</Description>
-		<ProjectGuid>D42DD1C7-055E-4DE9-B674-E0BB113A0996</ProjectGuid>
+		<RootNamespace>OLT.Core</RootNamespace>
 	</PropertyGroup>
-
+	
 	<ItemGroup>
-		<PackageReference Include="Microsoft.Extensions.Caching.Abstractions" Version="8.0.0" />
-		<PackageReference Include="StackExchange.Redis" Version="2.7.33" />
-		<PackageReference Include="StackExchange.Redis.Extensions.Core" Version="10.2.0" />
+		<PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="9.0.0" />
+		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="9.0.0" />
+		<PackageReference Include="Microsoft.Extensions.Options" Version="9.0.0" />		
 		<PackageReference Include="StackExchange.Redis.Extensions.Newtonsoft" Version="10.2.0" />
 	</ItemGroup>
 
 	<ItemGroup>
-		<ProjectReference Include="..\OLT.Extensions.Caching.Shared\OLT.Extensions.Caching.Shared.csproj" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <PackageReference Update="Microsoft.SourceLink.GitHub" Version="8.0.0" />
+	  <ProjectReference Include="..\OLT.Extensions.Caching.Memory\OLT.Extensions.Caching.Memory.csproj" />
 	</ItemGroup>
 
 </Project>
