@@ -1,11 +1,10 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-using StackExchange.Redis;
+﻿using Microsoft.Extensions.Caching.Memory;
+using Microsoft.Extensions.DependencyInjection;
 using StackExchange.Redis.Extensions.Core;
 using StackExchange.Redis.Extensions.Core.Abstractions;
 using StackExchange.Redis.Extensions.Core.Configuration;
 using StackExchange.Redis.Extensions.Core.Implementations;
 using StackExchange.Redis.Extensions.Newtonsoft;
-using System;
 
 namespace OLT.Core
 {
@@ -23,12 +22,12 @@ namespace OLT.Core
         /// <param name="cacheKeyPrefix">Keyprefix for namespacing cache entries</param>
         /// <param name="connectionString"> The string configuration to use for this <see cref="RedisConfiguration.ConnectionString"/>. ("localhost:6379,serviceName=application2,name=my-test-app2,defaultDatabase=3")</param>
         /// <returns><seealso cref="IServiceCollection"/></returns>
+        /// <exception cref="ArgumentNullException"></exception>
+        /// <exception cref="ArgumentException"></exception>
+        [Obsolete("Removing in 10.x, Move to FusionCache")]
         public static IServiceCollection AddOltCacheRedis(this IServiceCollection services, TimeSpan defaultAbsoluteExpiration, string cacheKeyPrefix, string connectionString)
         {
-            if (string.IsNullOrWhiteSpace(connectionString))
-            {
-                throw new ArgumentNullException(nameof(connectionString));
-            }
+            ArgumentNullException.ThrowIfNullOrWhiteSpace(connectionString);
             return AddOltCacheRedis(services, defaultAbsoluteExpiration, cacheKeyPrefix, new RedisConfiguration { ConnectionString = connectionString });
         }
 
@@ -44,36 +43,30 @@ namespace OLT.Core
         /// <param name="redisConfiguration"></param>
         /// <returns></returns>
         /// <exception cref="ArgumentNullException"></exception>
+        /// <exception cref="ArgumentException"></exception>
+        [Obsolete("Removing in 10.x, Move to FusionCache")]
         public static IServiceCollection AddOltCacheRedis(this IServiceCollection services, TimeSpan defaultAbsoluteExpiration, string cacheKeyPrefix, RedisConfiguration redisConfiguration)
         {
-            if (services == null)
-            {
-                throw new ArgumentNullException(nameof(services));
-            }
-
-            if (redisConfiguration == null)
-            {
-                throw new ArgumentNullException(nameof(redisConfiguration));
-            }
-
-            if (string.IsNullOrWhiteSpace(cacheKeyPrefix))
-            {
-                throw new ArgumentNullException(nameof(cacheKeyPrefix));
-            }
+            ArgumentNullException.ThrowIfNull(services);
+            ArgumentNullException.ThrowIfNull(redisConfiguration);
+            ArgumentNullException.ThrowIfNullOrWhiteSpace(cacheKeyPrefix);
 
             redisConfiguration.Name = redisConfiguration.Name ?? redisConfiguration.ConfigurationOptions.ClientName;
+            redisConfiguration.AllowAdmin = true;
 
-            services.AddSingleton((provider) => provider
-                .GetRequiredService<IRedisClientFactory>()
-                .GetDefaultRedisClient());
-
-            services.AddSingleton((provider) => provider
-                .GetRequiredService<IRedisClientFactory>()
-                .GetDefaultRedisClient()
-                .GetDefaultDatabase());
+            services.AddSingleton((provider) => provider.GetRequiredService<IRedisClientFactory>().GetDefaultRedisClient());
+            services.AddSingleton((provider) => provider.GetRequiredService<IRedisClientFactory>().GetDefaultRedisClient().GetDefaultDatabase());
+            
+            services
+                .AddMemoryCache(opt => new MemoryCacheEntryOptions().SetAbsoluteExpiration(defaultAbsoluteExpiration))
+                .AddSingleton<IOltMemoryCache, OltMemoryCache>();
 
             return services
-                .AddSingleton<IOltCacheService, OltRedisCache>()
+                .Configure<OltCacheOptions>(opt =>
+                {
+                    opt.DefaultAbsoluteExpiration = defaultAbsoluteExpiration;                    
+                })
+                .AddSingleton<IOltCacheService, OltRedisCache>()                
                 .AddSingleton(redisConfiguration)
                 .AddSingleton<IRedisClientFactory, RedisClientFactory>()
                 .AddSingleton<IRedisConnectionPoolManager, RedisConnectionPoolManager>()
