@@ -1,7 +1,6 @@
-﻿using System.Diagnostics;
-using System.Linq;
+﻿using System;
+using System.Diagnostics;
 using Microsoft.Extensions.Logging;
-using OLT.Constants;
 using OLT.Logging.Serilog;
 using OLT.Logging.Serilog.Enricher;
 using Serilog;
@@ -20,24 +19,29 @@ namespace OLT.Core
         /// <returns>Configuration object allowing method chaining.</returns>
         public static LoggerConfiguration WithOltEventType(this LoggerEnrichmentConfiguration loggerConfiguration)
         {
-            return loggerConfiguration
-                .With(new OltEventTypeEnricher());
+            return loggerConfiguration.With(new OltEventTypeEnricher());
         }
 
 
         /// <summary>
         /// Enrich log the Environment Name <see cref="OltSerilogConstants.Properties.Environment"/> and <see cref="OltSerilogConstants.Properties.DebuggerAttached"/>
         /// </summary>
-        /// <returns>Configuration object allowing method chaining.</returns>
+        /// <returns>Configuration object allowing method chaining.</returns>        
+        [Obsolete("Use to Serilog.Enrichers.Environment -> WithEnvironmentName")]
         public static LoggerConfiguration WithOltEnvironment(this LoggerEnrichmentConfiguration loggerConfiguration, string environmentName)
         {
+
             return loggerConfiguration
                 .WithProperty(OltSerilogConstants.Properties.Environment, environmentName)
+
                 // Used to filter out potentially bad data due debugging.
                 // Very useful when doing Seq dashboards and want to remove logs under debugging session.
-                .Enrich.WithProperty(OltSerilogConstants.Properties.DebuggerAttached, Debugger.IsAttached);
+                .Enrich.WithProperty(OltSerilogConstants.Properties.DebuggerAttached, Debugger.IsAttached)
+                ;
         }
 
+
+    
 
 
         /// <summary>
