@@ -17,15 +17,14 @@ namespace OLT.Extensions.DependencyInjection.Tests
             Assembly? nullRef = null;
             List<Assembly>? nullList = null;
 
-            Assert.Throws<ArgumentNullException>("services", () => OltServiceCollectionExtensions.AddOltInjection(null!));
-            Assert.Throws<ArgumentNullException>("baseAssembly", () => OltServiceCollectionExtensions.AddOltInjection(null!, nullRef!));
-            Assert.Throws<ArgumentNullException>("services", () => OltServiceCollectionExtensions.AddOltInjection(null!, nullList!));
-
-            Assert.Throws<ArgumentNullException>("baseAssembly", () => OltServiceCollectionExtensions.AddOltInjection(services, nullRef!));
+            Assert.Throws<ArgumentNullException>("services", () => OltServiceCollectionExtensions.AddOltInjection(null!, null));
+            Assert.Throws<ArgumentNullException>("baseAssembly", () => OltServiceCollectionExtensions.AddOltInjection(null!, nullRef!, null));
+            Assert.Throws<ArgumentNullException>("services", () => OltServiceCollectionExtensions.AddOltInjection(null!, nullList!, null));
+            Assert.Throws<ArgumentNullException>("baseAssembly", () => OltServiceCollectionExtensions.AddOltInjection(services, nullRef!, null));
 
             try
             {
-                OltServiceCollectionExtensions.AddOltInjection(services, nullList!);
+                OltServiceCollectionExtensions.AddOltInjection(services, nullList!, null);
                 Assert.True(true);
             }
             catch
@@ -38,8 +37,10 @@ namespace OLT.Extensions.DependencyInjection.Tests
         public void InjectionTest()
         {
             var services = new ServiceCollection();
-            OltServiceCollectionExtensions.AddOltInjection(services)
-                .AddScoped<IOltIdentity, TestIdentity>();
+            OltServiceCollectionExtensions.AddOltInjection(services, null)
+                .AddScoped<IOltIdentity, TestIdentity>()
+                .AddScoped<IOltDbAuditUser>(x => x.GetRequiredService<IOltIdentity>())
+                ;
             
             using (var provider = services.BuildServiceProvider())
             {
@@ -52,7 +53,8 @@ namespace OLT.Extensions.DependencyInjection.Tests
         {
             var services = new ServiceCollection();
             OltServiceCollectionExtensions.AddOltInjection(services, this.GetType().Assembly)
-                .AddScoped<IOltIdentity, TestIdentity>();
+                .AddScoped<IOltIdentity, TestIdentity>()
+                .AddScoped<IOltDbAuditUser>(x => x.GetRequiredService<IOltIdentity>());
 
             using (var provider = services.BuildServiceProvider())
             {
@@ -65,12 +67,33 @@ namespace OLT.Extensions.DependencyInjection.Tests
         {
             var services = new ServiceCollection();
             OltServiceCollectionExtensions.AddOltInjection(services, new List<Assembly>() { this.GetType().Assembly })
-                .AddScoped<IOltIdentity, TestIdentity>();
+                .AddScoped<IOltIdentity, TestIdentity>()
+                .AddScoped<IOltDbAuditUser>(x => x.GetRequiredService<IOltIdentity>());
 
             using (var provider = services.BuildServiceProvider())
             {
                 Assert.NotNull(provider.GetService<IOltDbAuditUser>());
             }
         }
+
+        [Fact]
+        public void Test_Scan()
+        {
+            var services = new ServiceCollection();
+            OltServiceCollectionExtensions
+                .Scan(services, opt =>
+                {
+                    opt.IncludeAssembly(this.GetType().Assembly);
+                    opt.DeepScan();
+                })
+                .AddScoped<IOltIdentity, TestIdentity>()
+                .AddScoped<IOltDbAuditUser>(x => x.GetRequiredService<IOltIdentity>());
+
+            using (var provider = services.BuildServiceProvider())
+            {
+                Assert.NotNull(provider.GetService<IOltDbAuditUser>());
+            }
+        }
+
     }
 }
